Move StyledComponents to SCSS Modules
