+++ src/components/grid-system/col/styles.module.scss
@@ -1,0 +1,9 @@
+.col {
+  flex-grow: 0;
+  flex-shrink: 0;
+  min-height: 1px;
+  position: relative;
+  left: 0;
+  right: 0;
+  max-width: 100%;
+}
