+++ src/components/grid-system/container/styles.module.scss
@@ -1,0 +1,7 @@
+.container {
+  box-sizing: border-box;
+  margin-left: auto;
+  margin-right: auto;
+  position: relative;
+  max-width: none;
+}
