+++ src/components/grid-system/flex/index.js
@@ -2,14 +2,13 @@ import React from "react"
 import PropTypes from "prop-types"
 import styled from "styled-components"
 
+import * as styles from "./styles.module.scss"
+
 const SFlex = styled.div`
   margin: ${props => (props.$fluid ? "0rem" : "0rem -1.4rem")};
-  display: flex;
   flex-wrap: ${props => (props.$nowrap ? "nowrap" : "wrap")};
   justify-content: ${props => props.$justify};
   align-items: ${props => props.$align};
-  flex-grow: 0;
-  flex-shrink: 0;
 `
 
 const Flex = ({ children, ...props }) => (
@@ -18,6 +17,7 @@ const Flex = ({ children, ...props }) => (
     $justify={props.justify}
     $align={props.align}
     $fluid={props.fluid ? 1 : 0}
+    className={styles.flex}
     {...props}
   >
     {children}
