+++ src/components/grid-system/flex/styles.module.scss
@@ -1,0 +1,5 @@
+.flex {
+  display: flex;
+  flex-grow: 0;
+  flex-shrink: 0;
+}
