+++ src/styles/global/_styles.scss
@@ -1,0 +1,149 @@
+@use "../vars" as *;
+
+html {
+  background: linear-gradient(150deg, var(--color-navy) 8%, var(--color-black) 100%);
+
+  font-size: 62.5%;
+  line-height: 1.65;
+  box-sizing: border-box;
+
+  color: var(--color-white);
+  font-family: var(--font-bai);
+  -ms-text-size-adjust: 100%;
+  -webkit-text-size-adjust: 100%;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+*,
+*:before,
+*:after {
+  box-sizing: inherit;
+}
+
+body {
+  font-size: 1.8rem;
+  font-weight: 400;
+}
+
+#site-wrapper {
+  min-height: 100vh;
+  display: grid;
+  grid-template-rows: [header] min-content [main] auto [footer] min-content;
+}
+
+a {
+  color: blue;
+
+  &:hover {
+    text-decoration: underline;
+  }
+}
+
+img,
+svg {
+  display: inline-block;
+  vertical-align: middle;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  margin: 0;
+  font-family: var(--font-exo);
+  font-style: italic;
+}
+
+hgroup {
+  margin-bottom: 1.25em;
+}
+
+h1 {
+  font-size: 5.6rem;
+  text-align: center;
+}
+
+h2 {
+  font-size: 2rem;
+  text-align: center;
+}
+
+h3 {
+  font-size: 3.6rem;
+  text-align: left;
+}
+
+h4 {
+  font-size: 2.2rem;
+}
+
+h5 {
+  font-size: 1.5rem;
+  text-align: center;
+}
+
+h6 {
+  font-size: 1.3rem;
+}
+
+p {
+  margin: 0;
+  margin-bottom: 2.8rem;
+}
+
+*.text-xs,
+*.text-xs * {
+  font-size: 1.3rem;
+}
+
+*.text-sm,
+*.text-sm * {
+  font-size: 1.7rem;
+}
+
+*.text-md,
+*.text-md * {
+  font-size: 1.8rem;
+}
+
+.highlighted p:first-of-type,
+*.text-lg,
+*.text-lg * {
+  font-size: 2.9rem;
+}
+
+*.text-light,
+*.text-light * {
+  color: var(--color-gray-100);
+}
+
+*.text-center,
+*.text-center * {
+  text-align: center;
+}
+
+.background {
+  background-color: black;
+  color: white;
+  box-shadow: 0 0.4rem 1rem rgba(0, 0, 20, 0.2);
+}
+
+@media screen and (max-width: 720px) {
+  body,
+  .text-md,
+  .text-sm {
+    font-size: 1.6rem;
+  }
+
+  h1 {
+    font-size: 4.8rem;
+  }
+
+  .highlighted p:first-of-type,
+  *.text-lg,
+  *.text-lg * {
+    font-size: 2.3rem;
+  }
+}
