+++ src/styles/global/index.scss
@@ -1,1 +1,3 @@
 @use "../vars" as *;
+@use "normalize" as *;
+@use "styles" as *;
