+++ src/styles/mixins/_spacing.scss
@@ -8,3 +8,13 @@
     margin-bottom: 6rem;
   }
 }
+
+@mixin space-sm {
+  margin-top: 4rem;
+  margin-bottom: 4rem;
+
+  :first-of-type {
+    margin-top: 2rem;
+    margin-bottom: 2rem;
+  }
+}
