+++ src/styles/vars/_fonts.scss
@@ -1,4 +1,4 @@
 :root {
-  --font-bai: "Bai Jamjuree', 'Lucida Grande', sans-serif";
-  --font-exo: "Exo, 'Lucida Grande', sans-serif";
+  --font-bai: "Bai Jamjuree", "Lucida Grande", sans-serif;
+  --font-exo: "Exo", "Lucida Grande", sans-serif;
 }
